LayoutTests/transforms/3d/hit-testing/overlapping-layers-hit-test-expected.txt

Element at 10, 100 has id "container-nopreserve": PASS
Element at 20, 100 has id "target2": PASS
Element at 50, 100 has id "target1": PASS
Element at 80, 100 has id "target1": PASS
Element at 100, 100 has id "target1": PASS
Element at 120, 100 has id "target1": PASS
Element at 180, 100 has id "target1": PASS
Element at 190, 100 has id "container-nopreserve": PASS
Element at 10, 250 has id "container-preserve": PASS
Element at 20, 250 has id "target4": PASS
Element at 80, 250 has id "target4": PASS
Element at 100, 250 has id "target4": PASS
Element at 120, 250 has id "target3": PASS
Element at 180, 250 has id "target3": PASS
Element at 190, 250 has id "container-preserve": PASS
Element at 10, 550 has id "container-preserve": PASS
Element at 20, 550 has id "target5": PASS
Element at 50, 550 has id "target6": PASS
Element at 80, 550 has id "target6": PASS
Element at 100, 550 has id "target6": PASS
Element at 120, 550 has id "target6": PASS
Element at 180, 550 has id "target6": PASS
Element at 190, 550 has id "container-preserve": PASS
